+++ plugins/template/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bs-plugins/template
 
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies [24cc99e]
+  - @bs-core/shell@1.9.1
+
 ## 0.0.13
 
 ### Patch Changes